+++ player.h
@@ -19,7 +19,7 @@ public :
 	player(const player& other);
 	player& operator=(const player& other);
 
-	int return_damage_of_attack(int a), return_dm_of_special();
+	float return_damage_of_attack(int a), return_dm_of_special();
 	void draw_text_call(int a,bool facing);
 	bool allow_sp_attack();
 	void receiveDamage(int base_damage)override;
